+++ app/src/firebase/InitializeDatabase.ts
