+++ app/src/firebase/snaps/Snap.ts
