+++ app/src/firebase/users/UserProfile.ts
@@ -1,0 +1,4 @@
+export interface UserProfile {
+    isAdmin: boolean;
+    isSnapper: boolean;
+}
