+++ app/src/firebase/users/__mocks__/UserService.ts
@@ -1,7 +1,4 @@
-export interface ProfileData {
-    isAdmin: boolean;
-    isSnapper: boolean;
-}
+import { UserProfile } from "../UserProfile";
 
 export async function signIn() {
     return;
@@ -15,7 +12,7 @@ export function getCurrentUserName(): String {
     return "Henry";
 }
 
-export function onAuthStateChanged(cb: (p: ProfileData) => void) {
+export function onAuthStateChanged(cb: (p: UserProfile) => void) {
     const mockProfile = { isAdmin: true, isSnapper: true };
     cb(mockProfile);
 }
